requirements.txt
requests>=2.31.0
streamlit>=1.28.0
chromadb>=0.4.15
sentence-transformers>=2.3.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
apscheduler>=3.10.0
plotly>=5.17.0
huggingface-hub>=0.19.0
transformers>=4.35.0